refactor: 폴더, 파일명 일부 변경
